refactor(qa-run): rozbicie runnera macierzy na moduły
- Wydzielono odpowiedzialności run-matrix do modułów app, config, parsers, patterns, change-detection, execution, cache, reporting, session i shared.
- Zachowano cienki entrypoint CLI z re-eksportem parseConfig dla kompatybilności testów i istniejących użyć.
- Dodano testy jednostkowe dla parserów, normalizacji configu, snapshotów, sesji, cache, output collectorów, executorów i summary writerów.
- Dodano konfigurację ESLint analogiczną do sąsiedniego projektu oraz podpięto lint do package.json i macierzy qa-run.
- Zaktualizowano macierz QA, aby zmiany w runnerze, testach i konfiguracji ESLint uruchamiały lint oraz testy.
