+++ .agents/skills/qa-run/scripts/run-matrix/shared/hashing.mjs
@@ -1,0 +1,9 @@
+import {createHash} from "node:crypto";
+
+export function hashBuffer(buffer) {
+    return createHash("sha256").update(buffer).digest("hex");
+}
+
+export function hashJson(value) {
+    return createHash("sha256").update(JSON.stringify(value)).digest("hex");
+}
